+++ tests/epoch.lisp
@@ -21,7 +21,7 @@
                                     :resume nil
                                     :self-mod t)))
     (is (getf out :flagship))
-    (is (equal "3.0.0" (getf out :version)))
+    (is (equal "3.1.0" (getf out :version)))
     (is (getf out :report))
     (is (getf out :complete))
     (is (eq :complete (getf (getf out :status) :status)))